services/orion-vllm/requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
